Updated Code logic updated code logic to not use an elseif statement for each senerio and made it so it only needs two elseif statements, one for win and one for loss.
The way the logic works is that rock,paper,scissors are assigned a value
1,2,3 respectively. The choice for the computer is a random number from
1 to 3 while the player choice is user inputted prompt. The computer
choice, player choice are sent to the play function as well as an array
and an outcome object. The way the play functions knows who won is that
for rock paper scissors, if one of the assign values is added by 2 and
the new value equals the value picked by the computer then the player
always wins. if the value after adding 2 does not equal the computer
then the player losses. A draw senerio is calculated if the inital value
of player and comp equal each other without adding 2 then they picked
the same item. The array is two sent into the function as a value of 3
and 2 are assigned for paper and scissors but the computer would only
have a value between 1 and 3. If 2 is added to the player choice it
would be 4 or 5 in some senerios so a array is sent that is wrapped and
if a value goes over 3 then it loops back to the start. the outcome obj
will let the prompt at the end know which choice the computer picked.